Rename `page.vue` to distinguish compile and runtime pages
