+++ site/frontend/src/pages/compare/compile/compile-page.vue
